* p4a-validation:
  help for debug.
  more examples for terapix tests.
  add preschedule gradient for terapix tests.
  validate changes due to use of pips_user_warning in properties.l
  sort added in points_to prettyprint to simplify validation and to improve readability
  A few more test cases to start testing the property ALIAS_ACROSS_TYPE
  Two modifications in results, mostly for Semantics: TOP-LEVEL is not longer printed when an external variable is used (this may not be perfect, function entity_more_or_less_minimal_name may have to be improved to keep TOP-LEVEL when the variable is not in the scope of the current function); the effects order must have changed... equations are now different, but equivalent

+++ packages/PIPS/validation/Terapix/average_power.c
@@ -7,7 +7,7 @@ typedef struct {
 void average_power(int Nth, int Nrg, int Nv, Cplfloat ptrin[Nth][Nrg][Nv],
         Cplfloat Pow[Nth]) {
 
-    double PP;
+    float PP;
     int th, v, rg;
 
     for (th=0; th<Nth; th++) {
